feat: lab 2 init

+++ mc_labs/mc_lab_02/lab_2/index/src/hardware/Led/Led.cpp
@@ -1,24 +1,25 @@
 #include <Arduino.h>
-#include "../include/Led.h"
 #include "Led.h"
+#include "Color.h"
+#include "../Pins.h"
 
 Led::Led(Pin pin, Color color) : pin(pin), color(color) {
-	pinMode(pin, OUTPUT);
+	pinMode(static_cast<uint8_t>(pin), OUTPUT);
 }
 
 void Led::on() {
-	digitalWrite(pin, HIGH);
+	digitalWrite(static_cast<uint8_t>(pin), HIGH);
 	state = true;
 }
 
 void Led::off() {
-	digitalWrite(pin, LOW);
+	digitalWrite(static_cast<uint8_t>(pin), LOW);
 	state = false;
 }
 
 void Led::toggle() {
 	state = !state;
-	digitalWrite(pin, state ? HIGH : LOW);
+	digitalWrite(static_cast<uint8_t>(pin), state ? HIGH : LOW);
 }
 
 bool Led::isOn() const {
